move from gui.py to main.py
